fix: BCLIBC_Config_fromPyObject static_cast fix

+++ py_ballisticcalc.exts/py_ballisticcalc_exts/src/py_bind.cpp
@@ -15,7 +15,7 @@ namespace bclibc
             PyFloat_AsDouble(PyObject_GetAttrString(config, "cZeroFindingAccuracy")),
             PyFloat_AsDouble(PyObject_GetAttrString(config, "cMinimumVelocity")),
             PyFloat_AsDouble(PyObject_GetAttrString(config, "cMaximumDrop")),
-            PyLong_AsLong(PyObject_GetAttrString(config, "cMaxIterations")),
+            (int)PyLong_AsLong(PyObject_GetAttrString(config, "cMaxIterations")),
             PyFloat_AsDouble(PyObject_GetAttrString(config, "cGravityConstant")),
             PyFloat_AsDouble(PyObject_GetAttrString(config, "cMinimumAltitude"))};
     };
